@@ -2,25 +2,31 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net9.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
     <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-    <AssemblyVersion>1.0.4</AssemblyVersion>
-    <FileVersion>1.0.4</FileVersion>
-	<Version>1.0.4</Version>
+    <AssemblyVersion>1.1.0</AssemblyVersion>
+    <FileVersion>1.1.0</FileVersion>
+	<Version>1.1.0</Version>
+    <PreserveCompilationContext>true</PreserveCompilationContext>
+	<RuntimeIdentifiers>win;osx;linux;</RuntimeIdentifiers>
 	<PackageReleaseNotes>Initial 1.4.4.9 support</PackageReleaseNotes>
   </PropertyGroup>
 	
   <ItemGroup>
-    <PackageReference Include="ModFramework.Modules.ClearScript" Version="1.1.7" />
-    <PackageReference Include="ModFramework.Modules.CSharp" Version="1.1.7" />
-    <PackageReference Include="ModFramework.Modules.Lua" Version="1.1.7" />
-    <PackageReference Include="MonoMod.RuntimeDetour.HookGen" Version="22.5.1.1" />
-    <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
-    <PackageReference Include="OTAPI.Upcoming" Version="3.1.20" />
-    <PackageReference Include="System.Security.Permissions" Version="6.0.0" />
-    <PackageReference Include="Steamworks.NET" Version="20.1.0" />
+	  <PackageReference Include="ModFramework" Version="1.1.13" />
+	  <PackageReference Include="ModFramework.Modules.ClearScript" Version="1.1.13" />
+	  <PackageReference Include="ModFramework.Modules.CSharp" Version="1.1.13" />
+	  <PackageReference Include="ModFramework.Modules.Lua" Version="1.1.13" />
+	  <PackageReference Include="Steamworks.NET" Version="2024.8.0" />
+	  <PackageReference Include="MonoMod" Version="22.7.31.1" />
+	  <PackageReference Include="MonoMod.RuntimeDetour.HookGen" Version="22.7.31.1" />
+	  <PackageReference Include="MonoMod.RuntimeDetour" Version="25.2.2" />
+	  <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
+	  <PackageReference Include="OTAPI.Upcoming" Version="3.2.4" />
+	  <PackageReference Include="System.Private.Uri" Version="4.3.2" />
+	  <PackageReference Include="System.Security.Permissions" Version="9.0.0" />
   </ItemGroup>
 	
   <ItemGroup>
